finished handlers for categories, transactions and budgets

+++ internal/models/budget.go
@@ -1,11 +1,9 @@
-// internal/models/budget.go
-
 package models
 
 type Budget struct {
-	ID     int    `json:"id"`
-	Name   string `json:"name"`
-	Amount int    `json:"amount"`
+	ID     int     `json:"id"`
+	Name   string  `json:"name"`
+	Amount float64 `json:"amount"`
 }
 
 // Add more models as needed
